requirements.txt
django==3.2.7
djangorestframework==3.13.1
django-cors-headers==3.13.0
gunicorn
django-environ==0.8.1
psycopg2
gunicorn==20.1.0
Pillow==9.0.1
drf-yasg
python-dotenv
django-modeltranslation
